Add only_a_flesh_wound_percent parameter.
If the board's state is identical to its state a set number of steps
ago, we assume the board has essentially become static, so we introduce
a `only_a_flesh_wound_percent` probaility for a cell to remain alive
even if it should have died.  We will keep increasing this probability
as long as we still seem to be stuck in this pattern.  Once we break out
of the pattern, we set that probabiliity back to 0.

Also, we created a cell_updater_generator() function that accepts a
only_a_flesh_wound_percent parameter and creates a cell_updater()
function that can be used that includes that paramter.